@@ -13,45 +13,47 @@
     "test-watch": "mocha --opts test/mocha.opts --watch test/tests.js"
   },
   "dependencies": {
-    "body-parser": "^1.15.2",
-    "express": "^4.14.0",
-    "lowdb": "^0.14.0"
+    "body-parser": "^1.17.1",
+    "express": "^4.15.2",
+    "lowdb": "^0.16.2"
   },
   "devDependencies": {
-    "autoprefixer": "^6.5.1",
-    "babel-core": "^6.18.0",
-    "babel-plugin-transform-object-rest-spread": "^6.16.0",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-preset-react": "^6.16.0",
+    "autoprefixer": "^6.7.7",
+    "babel-core": "^6.24.1",
+    "babel-plugin-transform-object-rest-spread": "^6.23.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
     "babelify": "^7.3.0",
     "browserify": "^13.1.1",
     "chai": "^3.5.0",
-    "csswring": "^5.1.0",
+    "create-react-class": "^15.5.2",
+    "csswring": "^5.1.1",
     "del": "^2.2.2",
-    "enzyme": "^2.5.1",
-    "eslint": "^3.12.2",
-    "eslint-plugin-react": "^6.8.0",
+    "enzyme": "^2.8.2",
+    "eslint": "^3.19.0",
+    "eslint-plugin-react": "^6.10.3",
     "flexboxgrid": "^6.3.1",
     "gulp": "^3.9.1",
-    "gulp-concat": "^2.6.0",
-    "gulp-postcss": "^6.2.0",
+    "gulp-concat": "^2.6.1",
+    "gulp-postcss": "^6.4.0",
     "gulp-rename": "^1.2.2",
-    "gulp-uglify": "^2.0.0",
-    "jquery": "^3.1.1",
-    "jsdom": "^9.8.3",
-    "karma": "^1.3.0",
-    "karma-mocha": "^1.2.0",
-    "mocha": "^3.1.2",
-    "normalize.css": "^5.0.0",
-    "postcss-cssnext": "^2.8.0",
-    "react": "^15.3.2",
-    "react-addons-css-transition-group": "^15.3.2",
-    "react-addons-test-utils": "^15.3.2",
-    "react-dom": "^15.3.2",
+    "gulp-uglify": "^2.1.2",
+    "jquery": "^3.2.1",
+    "jsdom": "^9.12.0",
+    "karma": "^1.6.0",
+    "karma-mocha": "^1.3.0",
+    "mocha": "^3.2.0",
+    "normalize.css": "^6.0.0",
+    "postcss-cssnext": "^2.10.0",
+    "prop-types": "^15.5.8",
+    "react": "^15.5.4",
+    "react-addons-css-transition-group": "^15.5.2",
+    "react-addons-test-utils": "^15.5.1",
+    "react-dom": "^15.5.4",
     "run-sequence": "^1.2.2",
     "testdom": "^2.0.0",
     "vinyl-buffer": "^1.0.0",
     "vinyl-source-stream": "^1.1.0",
-    "watchify": "^3.7.0"
+    "watchify": "^3.9.0"
   }
 }
